🌱 Split KustomizationDrillDown.tsx into orchestrator, parts, and hook
Extracts KustomizationDrillDown.tsx (546 lines, 12 hook calls)
following the established X.tsx + X.parts.tsx + useX.ts drilldown-
split pattern.

- useKustomizationDrillDown.ts — all state, the kubectl fetch effect
  for applied resources/conditions, and the handleDiagnose callback
- KustomizationDrillDown.parts.tsx — pure presentational
  sub-components (header, tab bar, overview/resources/conditions/ai
  tab panels) plus the getStatusStyle helper
- KustomizationDrillDown.tsx — thin orchestrator (115 lines, 2 hook
  calls)

No behaviour change — pure extraction. Drill-down registration and
useDrillDown() props unchanged.
